Make client.Connect obligatory
Make NewClient not automatically start connecting in the background
any more and make Connect obligatory instead.
When autoconnect is enabled NewClient should initialize the client
with autoconnect set to deactivated initially.
Change the API of the client by requiring a context parameter in
Connect which is used to make Connect cancelable.
Make Connect call tryAutoconnect when autoconnect is enabled.
Make Request, Signal and RestoreSession reactivate temporarily
disabled autoconnect.
Update tests according to the changes in the API.

fixes #2